Rename Assets RP comp to GetAssets
